% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minosse.plot.R
\name{minosse.plot}
\alias{minosse.plot}
\title{A function to plot the MInOSSE results}
\usage{
minosse.plot(minosse_res,th_num=3,pcex=1.5,cont_pol=NULL,title=NULL)
}
\arguments{
\item{minosse_res}{the output of minosse.target function or the minosse_res element list in the minosse function output.}

\item{th_num}{Numeric. The Regression Kriging prediction map binarization threshold value index as reported in the optimal.thresholds function in the PresenceAbsence package. Default 3 = MaxSens+Spec,    maximizes (sensitivity+specificity)/2.}

\item{pcex}{Numeric. The graphic parameter indicating the size of points showing target species fossil localities. Default 1.5.}

\item{cont_pol}{The spatial polygon object showing the land or sea contour on the maps. Default NULL.}

\item{title}{The title on the top of the plot. If NULL, then the name of the species is used. Default NULL}
}
\value{
A plot showing the occurrence probability (on the left) and the threshold-based gepgraphic range (on the right) maps of target species along with its fossil localities.
}
\description{
This function provides two maps: one with a species occurrence probability and another with the binarized geographic range
}
\details{
When running this function for the first time in your workspece it is necessary to firstly load ggplo2.
}
\examples{
  \donttest{

  library(raster)
  data(lgm)
  raster(system.file("exdata/prediction_ground.gri", package="PaleoCore"))->prediction_ground

  mam<-minosse(dat=lgm,species.name="Mammuthus_primigenius",domain="land",
  prediction.ground=prediction_ground,crop.by.mcp=FALSE,th_num=3,
  coc_by="locality",n.sims=10,projection="laea",lon_0 = 85,
  lat_0 = 45,seed=625)

  library(ggplot2)
  minosse.plot(mam)

  }
}
\author{
Francesco Carotenuto, francesco.carotenuto@unina.it
}
